+++ include/rocksdb/listener.h
@@ -21,23 +21,36 @@ class DB;
 class Status;
 struct CompactionJobStats;
 
-struct TableFileCreationInfo {
-  TableFileCreationInfo() = default;
-  explicit TableFileCreationInfo(TableProperties&& prop) :
-      table_properties(prop) {}
+enum class TableFileCreationReason {
+  kFlush,
+  kCompaction,
+  kRecovery,
+};
+
+struct TableFileCreationBriefInfo {
   // the name of the database where the file was created
   std::string db_name;
   // the name of the column family where the file was created.
   std::string cf_name;
   // the path to the created file.
   std::string file_path;
-  // the size of the file.
-  uint64_t file_size;
   // the id of the job (which could be flush or compaction) that
   // created the file.
   int job_id;
+  // reason of creating the table.
+  TableFileCreationReason reason;
+};
+
+struct TableFileCreationInfo : public TableFileCreationBriefInfo {
+  TableFileCreationInfo() = default;
+  explicit TableFileCreationInfo(TableProperties&& prop)
+      : table_properties(prop) {}
+  // the size of the file.
+  uint64_t file_size;
   // Detailed properties of the created file.
   TableProperties table_properties;
+  // The status indicating whether the creation was successful or not.
+  Status status;
 };
 
 enum class CompactionReason {
@@ -212,10 +225,24 @@ class EventListener {
   // on file creations and deletions is suggested to implement
   // OnFlushCompleted and OnCompactionCompleted.
   //
+  // Historically it will only be called if the file is successfully created.
+  // Now it will also be called on failure case. User can check info.status
+  // to see if it succeeded or not.
+  //
   // Note that if applications would like to use the passed reference
   // outside this function call, they should make copies from these
   // returned value.
   virtual void OnTableFileCreated(const TableFileCreationInfo& /*info*/) {}
+
+  // A call-back function for RocksDB which will be called before
+  // a SST file is being created. It will follow by OnTableFileCreated after
+  // the creation finishes.
+  //
+  // Note that if applications would like to use the passed reference
+  // outside this function call, they should make copies from these
+  // returned value.
+  virtual void OnTableFileCreationStarted(
+      const TableFileCreationBriefInfo& /*info*/) {}
 
   virtual ~EventListener() {}
 };